+++ src/khmerthings/__init__.py
@@ -2,14 +2,15 @@
 
 from khmerthings.clusters import segment_clusters
 from khmerthings.counter import WordCount, analyze, count_words
-from khmerthings.lexicon import Lexicon, default_lexicon
+from khmerthings.lexicon import WORD_SOURCES, Lexicon, default_lexicon, load_lexicon
 from khmerthings.segmenter import break_words, mark_boundaries
 from khmerthings.sorting import khmer_sort_key, sort_lines
 from khmerthings.tokenizer import Token, TokenType, tokenize
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
+    "WORD_SOURCES",
     "Lexicon",
     "Token",
     "TokenType",
@@ -20,6 +21,7 @@ __all__ = [
     "count_words",
     "default_lexicon",
     "khmer_sort_key",
+    "load_lexicon",
     "mark_boundaries",
     "segment_clusters",
     "sort_lines",